Add mockprovider plugin
